requirements.txt
datasets==2.13.1
keybert==0.7.0
llama_index==0.12.28
nest_asyncio==1.6.0
numpy==1.24.4
opencv_python_headless==4.11.0.86
pdf2image==1.17.0
peft==0.10.0
PyStemmer==2.2.0.3
PyStemmer==2.2.0.3
pytesseract==0.3.13
python-dotenv==1.1.0
scikit_learn==1.6.1
sentence_transformers==2.7.0
spacy==3.8.5
torch==2.6.0
transformers==4.40.2
